Constant.java     - Endpoint do serviço de serial structure. @#$ MD_Product_Serial_Structure.kt     - Compatibilidade do tipo com a tabela de produto serial. @#$ T_MD_Product_Serial_Structure.kt T_MD_Product_Serial_Structure_Env.kt T_MD_Product_Serial_Structure_Rec.kt     - Modelo de Envio e Respota. @#$ WBR_Product_Serial_Structure.java WS_Product_Serial_Structure.kt AndroidManifest.xml     - Serviço de consulta de estrutura de serial.     - Necessário definir fluxos alternativos. Ex: Serial não encontrado. @#$ Act020_Main.java Act020_Main_Presenter.java Act020_Main_Presenter_Impl.java     - Chamada do Serviço. @#$

+++ app/src/main/java/com/namoadigital/prj001/model/MD_Product_Serial_Structure.kt
@@ -2,7 +2,7 @@ package com.namoadigital.prj001.model
 
 class MD_Product_Serial_Structure(
     val customer_code: Long,
-    val product_code: Int,
+    val product_code: Long,
     val serial_code: Int,
     val has_item_check: Int,
     val scn_item_check: Int?,
